myApp

+++ app/src/main/java/com/example/forecastapp/modal/Coord.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Coord (
 
-  @SerializedName("lat" ) var lat : Int? = null,
-  @SerializedName("lon" ) var lon : Int? = null
+  @SerializedName("lat" ) var lat : Double? = null,
+  @SerializedName("lon" ) var lon : Double? = null
 
 )
