+++ app/src/main/java/com/example/forecastapp/modal/WeatherList.kt
@@ -12,7 +12,7 @@ data class WeatherList (
   @SerializedName("clouds"     ) var clouds     : Clouds?            = Clouds(),
   @SerializedName("wind"       ) var wind       : Wind?              = Wind(),
   @SerializedName("visibility" ) var visibility : Int?               = null,
-  @SerializedName("pop"        ) var pop        : Int?               = null,
+  @SerializedName("pop"        ) var pop        : Double?               = null,
   @SerializedName("sys"        ) var sys        : Sys?               = Sys(),
   @SerializedName("dt_txt"     ) var dtTxt      : String?            = null
